Simplify by removing useless stuff
The map() has already the index as second argument. Just use it. And
because the order for the arguments are the same between map() and the
render method, we can just pass the method reference instead of a
lambda.

Please note that the third argument of map is the ref on the array.
renderNotification hasn't a third method then it's not a problem.