Add scrolling restoration with infinite list (even variable)
But the accuracy is not 100% because it's just impossible to known
the exact list height when all items are variables. Only a render of all
items can fix this problem and it doesn't make sense. But here the trick
is to save and restore the internal ReactList cache. With this cache,
it's possible to provide something good for restoring the scroll
position. When the list is restored, it's possible that the new position
is one item lower that before. The second reason is that the ReactList
scrolling is not working with the pixel resolution, but it scrolls
according to an index.