TableCell: fix some warnings when rich text is provided to the cell
The table cell text properties can be used to inject some glyph or text
stylings in the cell.

Actually this props cannot be properly checked, so i use any
for skipping warning.

I tryed to refactor TableCell, but table use this trick with
the table rows schema, for injecting glyph or styling per row.

We can refactor this if we change the table schema/API